split out keyword compiler
